# packages/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files/clean": {
      "post": {
        "operationId": "cleanFiles",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleanRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Clean the given files: remove matching torrents and delete orphans.",
        "tags": [
          "files"
        ]
      }
    }
  },
  "info": {
    "title": "Cleanarr API",
    "description": "Scan for unused downloads and clean matching qBittorrent torrents.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
...
      "CleanRequestDto": {
        "type": "object",
        "properties": {
          "files": {
            "description": "Absolute paths of the unused files to clean.",
            "example": [
              "/mnt/Nas/Media/Movies/.downloads/Movie.2024.mkv"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deleteFiles": {
            "type": "boolean",
            "default": true,
            "description": "When true (default) the downloaded data is removed from disk along with the torrent. Set to false to only remove the torrent entry."
          }
        },
        "required": [
          "files"
        ]
      },
      "CleanupResultItemDto": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "removedTorrents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deletedFromDisk": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "removedTorrents",
          "deletedFromDisk"
        ]
      },
      "CleanupSummaryDto": {
        "type": "object",
        "properties": {
          "requested": {
            "type": "number"
          },
          "cleaned": {
            "type": "number"
          },
          "failed": {
            "type": "number"
          },
          "removedTorrentHashes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CleanupResultItemDto"
            }
          }
        },
        "required": [
          "requested",
          "cleaned",
          "failed",
          "removedTorrentHashes",
          "items"
        ]
      }
    }
  }
}